requirements.txt
python-telegram-bot==20.7
openai==1.3.7
python-dotenv==1.0.0
elevenlabs==0.2.26
requests==2.31.0
